move config files and remove flag
