+++ cmd/hgctl/config/testdata/config/output/out.bootstrap.yaml
