+++ cmd/hgctl/config/testdata/config/output/out.endpoints.yaml
