+++ pkg/cmd/hgctl/config_listener.go
@@ -17,7 +17,7 @@ package hgctl
 import (
 	"fmt"
 
-	"github.com/alibaba/higress/pkg/config"
+	"github.com/alibaba/higress/cmd/hgctl/config"
 	"github.com/spf13/cobra"
 	cmdutil "k8s.io/kubectl/pkg/cmd/util"
 )
